feat(devtools): Modulariza llm_interact.py e adiciona script de tarefa resolve-ac (#57)
Este commit inicia a refatoração do script `scripts/llm_interact.py` para uma arquitetura modular, conforme definido na Issue #57.

Principais alterações:
- Criado o pacote `scripts/llm_core/` contendo submódulos para funcionalidades compartilhadas (`api_client.py`, `args.py`, `config.py`, `context.py`, `io_utils.py`, `prompts.py`, `utils.py`).
- Migrada a tarefa `resolve-ac` para um script dedicado `scripts/tasks/llm_task_resolve_ac.py`, que agora utiliza as funcionalidades do `llm_core`.
- O script `scripts/llm_interact.py` foi atualizado para atuar como um dispatcher, invocando os scripts de tarefa. Uma cópia do script original foi mantida como `scripts/llm_interact_copy.py` para referência durante a migração.
- Adicionados arquivos `.gitignore` aos novoAs diretórios de scripts para excluir caches Python (`__pycache__`, `*.pyc`).

Esta modularização visa melhorar a organização, manutenibilidade e testabilidade do código de interação com LLMs.
Atende aos Critérios de Aceite 1 (AC1), 2 (AC2) e 3 (AC3) da Issue #57.

Refs #57
